Move PGPSignature utils.
